react-langgraph/pyproject.toml
[project]
name = "react-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
